+++ src/main/kotlin/io/foxcapades/lib/pdk/UByteDeque.kt
@@ -176,99 +176,18 @@ class UByteDeque : PrimitiveDeque<UByte, UByteArray> {
 
   // region Remove
 
-  /**
-   * Removes the first element of this deque.
-   *
-   * If this deque was empty, this method does nothing.
-   *
-   * This method differs from [pop] in 2 ways:
-   *
-   * 1. This method does not return the removed value.
-   * 2. This method does not throw an exception if the deque was empty on method
-   *    call.
-   */
-  fun removeHead() {
-    if (!isEmpty) {
-      realHead = incremented(realHead)
-      size--
+  override fun removeHead(count: Int) {
+    when {
+      count < 0     -> throw IllegalArgumentException()
+      isEmpty       -> {}
+      count == 0    -> {}
+      count >= size -> clear()
+      else          -> {
+        realHead = internalIndex(realHead + count)
+        size -= count
+      }
     }
   }
-
-  /**
-   * Removes the first element of this deque.
-   *
-   * Alias of [removeHead]
-   *
-   * If this deque was empty, this method does nothing.
-   *
-   * This method differs from [pop] in 2 ways:
-   *
-   * 1. This method does not return the removed value.
-   * 2. This method does not throw an exception if the deque was empty on method
-   *    call.
-   */
-  inline fun removeFirst() = removeHead()
-
-  /**
-   * Removes the first element of this deque.
-   *
-   * Alias of [removeHead]
-   *
-   * If this deque was empty, this method does nothing.
-   *
-   * This method differs from [pop] in 2 ways:
-   *
-   * 1. This method does not return the removed value.
-   * 2. This method does not throw an exception if the deque was empty on method
-   *    call.
-   */
-  inline fun removeFront() = removeHead()
-
-  /**
-   * Removes the first element of this deque.
-   *
-   * Alias of [removeHead]
-   *
-   * If this deque was empty, this method does nothing.
-   *
-   * This method differs from [pop] in 2 ways:
-   *
-   * 1. This method does not return the removed value.
-   * 2. This method does not throw an exception if the deque was empty on method
-   *    call.
-   */
-  inline fun deleteHead() = removeHead()
-
-  /**
-   * Removes the first element of this deque.
-   *
-   * Alias of [removeHead]
-   *
-   * If this deque was empty, this method does nothing.
-   *
-   * This method differs from [pop] in 2 ways:
-   *
-   * 1. This method does not return the removed value.
-   * 2. This method does not throw an exception if the deque was empty on method
-   *    call.
-   */
-  inline fun deleteFirst() = removeHead()
-
-  /**
-   * Removes the first element of this deque.
-   *
-   * Alias of [removeHead]
-   *
-   * If this deque was empty, this method does nothing.
-   *
-   * This method differs from [pop] in 2 ways:
-   *
-   * 1. This method does not return the removed value.
-   * 2. This method does not throw an exception if the deque was empty on method
-   *    call.
-   */
-  inline fun deleteFront() = removeHead()
-
 
   // endregion Remove
 
@@ -443,98 +362,15 @@ class UByteDeque : PrimitiveDeque<UByte, UByteArray> {
 
   // region Remove
 
-  /**
-   * Removes the last element of this deque.
-   *
-   * If this deque was empty, this method does nothing.
-   *
-   * This method differs from [popTail] in 2 ways:
-   *
-   * 1. This method does not return the removed value.
-   * 2. This method does not throw an exception if the deque was empty on method
-   *    call.
-   */
-  fun removeTail() {
-    if (!isEmpty) {
-      size--
+  override fun removeTail(count: Int) {
+    when {
+      count  < 0    ->  throw IllegalArgumentException()
+      count == 0    -> {}
+      isEmpty       -> {}
+      count >= size -> clear()
+      else          -> size -= count
     }
   }
-
-  /**
-   * Removes the last element of this deque.
-   *
-   * Alias of [removeTail]
-   *
-   * If this deque was empty, this method does nothing.
-   *
-   * This method differs from [popTail] in 2 ways:
-   *
-   * 1. This method does not return the removed value.
-   * 2. This method does not throw an exception if the deque was empty on method
-   *    call.
-   */
-  inline fun removeLast() = removeTail()
-
-  /**
-   * Removes the last element of this deque.
-   *
-   * Alias of [removeTail]
-   *
-   * If this deque was empty, this method does nothing.
-   *
-   * This method differs from [popTail] in 2 ways:
-   *
-   * 1. This method does not return the removed value.
-   * 2. This method does not throw an exception if the deque was empty on method
-   *    call.
-   */
-  inline fun removeBack() = removeTail()
-
-  /**
-   * Removes the last element of this deque.
-   *
-   * Alias of [removeTail]
-   *
-   * If this deque was empty, this method does nothing.
-   *
-   * This method differs from [popTail] in 2 ways:
-   *
-   * 1. This method does not return the removed value.
-   * 2. This method does not throw an exception if the deque was empty on method
-   *    call.
-   */
-  inline fun deleteTail() = removeTail()
-
-  /**
-   * Removes the last element of this deque.
-   *
-   * Alias of [removeTail]
-   *
-   * If this deque was empty, this method does nothing.
-   *
-   * This method differs from [popTail] in 2 ways:
-   *
-   * 1. This method does not return the removed value.
-   * 2. This method does not throw an exception if the deque was empty on method
-   *    call.
-   */
-  inline fun deleteLast() = removeTail()
-
-  /**
-   * Removes the last element of this deque.
-   *
-   * Alias of [removeTail]
-   *
-   * If this deque was empty, this method does nothing.
-   *
-   * This method differs from [popTail] in 2 ways:
-   *
-   * 1. This method does not return the removed value.
-   * 2. This method does not throw an exception if the deque was empty on method
-   *    call.
-   */
-  inline fun deleteBack() = removeTail()
-
 
   //////////////////////////////////////////////////////////////////////////////
   // endregion Remove
